backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
python-multipart==0.0.12
httpx==0.27.2
beautifulsoup4==4.12.3
anthropic==0.40.0
python-dotenv==1.0.1
aiofiles==24.1.0
lxml==5.3.0
psycopg2-binary==2.9.9
